+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ### `@liveblocks/react`
 
+- Fix a bug where under some conditions threads could end up without comments.
 - Fix a bug where notifications associated to deleted threads would not be
   deleted.
 - Fix a bug where subsequent optimistic updates to the same inbox notification